Refactor Assistant UseCases - User and CTOR
