+++ Application Layer/UserServices/UserServices.BusinessLayer/UseCase/AssistantRole.cs
@@ -11,98 +11,6 @@ using OnlineServices.Common.Exceptions;
 
 namespace RegistrationServices.BusinessLayer.UseCase
 {
-    public partial class AssistantRole : IRSAssistantRoleUser
-    {
-        private readonly IRSUnitOfWork iRSUnitOfWork;
-
-        public AssistantRole(IRSUnitOfWork iRSUnitOfWork)
-        {
-            this.iRSUnitOfWork = iRSUnitOfWork ?? throw new System.ArgumentNullException(nameof(iRSUnitOfWork));
-        }
-
-        public AssistantRole()
-        {
-
-        }
-
-        public bool AddUser(UserTO userTO)
-        {
-            if (userTO is null)
-                throw new LoggedException(new ArgumentNullException(nameof(userTO)));
-
-            if (userTO.Id != 0)
-                throw new Exception("Existing user");
-
-            userTO.Name.IsNullOrWhiteSpace("Missing User Name.");
-
-            try
-            {
-                iRSUnitOfWork.UserRepository.Add(userTO.ToDomain().ToTransfertObject());
-
-                return true;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-        }
-
-        public bool UpdateUser(UserTO userTO)
-        {
-            if (userTO is null)
-                throw new ArgumentNullException((nameof(userTO)));
-
-            if (userTO.Id == 0)
-                throw new Exception("User does not exist");
-
-            userTO.Name.IsNullOrWhiteSpace("Missing User Name");
-
-            try
-            {
-                iRSUnitOfWork.UserRepository.Update(userTO.ToDomain().ToTransfertObject());
-                return true;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-        }
-
-        public bool RemoveUser(UserTO userTO)
-        {
-            if (userTO is null)
-                throw new ArgumentNullException(nameof(userTO));
-
-            if (userTO.Id == 0)
-                throw new Exception("User does not exist");
-
-            try
-            {
-                iRSUnitOfWork.UserRepository.Remove(userTO.ToDomain().ToTransfertObject());
-                
-                return true;
-            }
-            catch (Exception)
-            {
-                throw;
-            }
-
-        }
-
-        public List<UserTO> GetUsers()
-        {
-            return iRSUnitOfWork.UserRepository.GetAll().Select(x => x.ToDomain().ToTransfertObject()).ToList();
-        }
-
-        public UserTO GetUserById(int id)
-        {
-            return iRSUnitOfWork.UserRepository.GetById(id);
-        }
-    }
-
-
-    //===================================================================================================================================
-
     public partial class AssistantRole : IRSAssistantRoleCourse
     {
         public bool AddCourse(CourseTO courseTo)
